Move all actor code to a top-level actors/ directory.
This is in preparation for extraction to a new repository as an executable Go project.
