+++ src/actors/builtin/account/account_actor.go
@@ -1,4 +1,4 @@
-package sysactors
+package account
 
 import (
 	ipld "github.com/filecoin-project/specs/libraries/ipld"
@@ -7,6 +7,8 @@ import (
 	vmr "github.com/filecoin-project/specs/systems/filecoin_vm/runtime"
 	exitcode "github.com/filecoin-project/specs/systems/filecoin_vm/runtime/exitcode"
 )
+
+type InvocOutput = vmr.InvocOutput
 
 func (a *AccountActorCode_I) Constructor(rt vmr.Runtime) InvocOutput {
 	// Nothing. intentionally left blank.
