+++ src/actors/builtin/cron/cron_actor.go
@@ -1,4 +1,4 @@
-package sysactors
+package cron
 
 import actor "github.com/filecoin-project/specs/systems/filecoin_vm/actor"
 import addr "github.com/filecoin-project/specs/systems/filecoin_vm/actor/address"
@@ -9,6 +9,9 @@ import vmr "github.com/filecoin-project/specs/systems/filecoin_vm/runtime"
 const (
 	Method_CronActor_EpochTick = actor.MethodPlaceholder + iota
 )
+
+type InvocOutput = vmr.InvocOutput
+type Runtime = vmr.Runtime
 
 func (a *CronActorCode_I) Constructor(rt vmr.Runtime) InvocOutput {
 	// Nothing. intentionally left blank.
