+++ src/actors/builtin/storage_power/storage_power_actor.id
@@ -1,4 +1,4 @@
-import actor_util "github.com/filecoin-project/specs/systems/filecoin_vm/actor_util"
+import actor_util "github.com/filecoin-project/specs/actors/util"
 import addr "github.com/filecoin-project/specs/systems/filecoin_vm/actor/address"
 import block "github.com/filecoin-project/specs/systems/filecoin_blockchain/struct/block"
 import actor "github.com/filecoin-project/specs/systems/filecoin_vm/actor"
