+++ src/actors/builtin/storage_power/storage_power_actor_boilerplate.go
@@ -1,18 +1,20 @@
-package storage_market
+package storage_power
 
 import (
+	actor_util "github.com/filecoin-project/specs/actors/util"
 	ipld "github.com/filecoin-project/specs/libraries/ipld"
 	actor "github.com/filecoin-project/specs/systems/filecoin_vm/actor"
-	actor_util "github.com/filecoin-project/specs/systems/filecoin_vm/actor_util"
 	vmr "github.com/filecoin-project/specs/systems/filecoin_vm/runtime"
 	util "github.com/filecoin-project/specs/util"
 )
 
 type BalanceTableHAMT = actor_util.BalanceTableHAMT
-type DealIDQueue = actor_util.DealIDQueue
+type SectorStorageWeightDesc = actor_util.SectorStorageWeightDesc
+type SectorTerminationType = actor_util.SectorTerminationType
+
+var SectorTerminationType_NormalExpiration = actor_util.SectorTerminationType_NormalExpiration
 
 var RT_MinerEntry_ValidateCaller_DetermineFundsLocation = vmr.RT_MinerEntry_ValidateCaller_DetermineFundsLocation
-var RT_ValidateImmediateCallerIsSignable = vmr.RT_ValidateImmediateCallerIsSignable
 
 ////////////////////////////////////////////////////////////////////////////////
 // Boilerplate
@@ -26,42 +28,32 @@ type InvocOutput = vmr.InvocOutput
 type Runtime = vmr.Runtime
 type Bytes = util.Bytes
 
+var Assert = util.Assert
 var IMPL_FINISH = util.IMPL_FINISH
 var IMPL_TODO = util.IMPL_TODO
+var PARAM_FINISH = util.PARAM_FINISH
+var TODO = util.TODO
 
-func (a *StorageMarketActorCode_I) State(rt Runtime) (vmr.ActorStateHandle, StorageMarketActorState) {
+func (a *StoragePowerActorCode_I) State(rt Runtime) (vmr.ActorStateHandle, StoragePowerActorState) {
 	h := rt.AcquireState()
-	var state StorageMarketActorState_I
 	stateCID := ipld.CID(h.Take())
+	var state StoragePowerActorState_I
 	if !rt.IpldGet(stateCID, &state) {
 		rt.AbortAPI("state not found")
 	}
 	return h, &state
 }
-func Release(rt Runtime, h vmr.ActorStateHandle, st StorageMarketActorState) {
+func Release(rt Runtime, h vmr.ActorStateHandle, st StoragePowerActorState) {
 	checkCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
 	h.Release(checkCID)
 }
-func UpdateRelease(rt Runtime, h vmr.ActorStateHandle, st StorageMarketActorState) {
+func UpdateRelease(rt Runtime, h vmr.ActorStateHandle, st StoragePowerActorState) {
 	newCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
 	h.UpdateRelease(newCID)
 }
-func (st *StorageMarketActorState_I) CID() ipld.CID {
-	IMPL_FINISH()
-	panic("")
+func (st *StoragePowerActorState_I) CID() ipld.CID {
+	panic("TODO")
 }
-
-func DealsAMT_Empty() DealsAMT {
-	IMPL_FINISH()
-	panic("")
-}
-
-func CachedDealIDsByPartyHAMT_Empty() CachedDealIDsByPartyHAMT {
-	IMPL_FINISH()
-	panic("")
-}
-
-func CachedExpirationsPendingHAMT_Empty() CachedExpirationsPendingHAMT {
-	IMPL_FINISH()
-	panic("")
+func DeserializeState(x Bytes) StoragePowerActorState {
+	panic("TODO")
 }
