--- src/systems/filecoin_blockchain/storage_power_consensus/consensus_fault_type_enum.go
@@ -1,10 +1,0 @@
-package storage_power_consensus
-
-type ConsensusFaultType int
-
-const (
-	UncommittedPowerFault ConsensusFaultType = 0
-	DoubleForkMiningFault ConsensusFaultType = 1
-	ParentGrindingFault   ConsensusFaultType = 2
-	TimeOffsetMiningFault ConsensusFaultType = 3
-)
