+++ src/systems/filecoin_blockchain/storage_power_consensus/expected_consensus.id
@@ -1,12 +1,13 @@
 import block "github.com/filecoin-project/specs/systems/filecoin_blockchain/struct/block"
 import chain "github.com/filecoin-project/specs/systems/filecoin_blockchain/struct/chain"
+import spowact "github.com/filecoin-project/specs/actors/builtin/storage_power"
 
 type ExpectedConsensus struct {
     expectedLeadersPerEpoch  UVarint
     expectedRewardPerEpoch   UVarint
 
     ComputeChainWeight(tipset chain.Tipset) block.ChainWeight
-    IsValidConsensusFault(faults ConsensusFaultType, blocks [block.Block]) bool
+    IsValidConsensusFault(faults spowact.ConsensusFaultType, blocks [block.Block]) bool
     IsWinningChallengeTicket(
         challengeTicket    util.Bytes
         sectorPower        block.StoragePower
