+++ src/systems/filecoin_blockchain/storage_power_consensus/storage_power_consensus_subsystem.go
@@ -3,6 +3,7 @@ package storage_power_consensus
 import (
 	"math"
 
+	spowact "github.com/filecoin-project/specs/actors/builtin/storage_power"
 	filcrypto "github.com/filecoin-project/specs/algorithms/crypto"
 	ipld "github.com/filecoin-project/specs/libraries/ipld"
 	block "github.com/filecoin-project/specs/systems/filecoin_blockchain/struct/block"
@@ -47,7 +48,7 @@ func (spc *StoragePowerConsensusSubsystem_I) IsWinningPartialTicket(stateTree st
 // TODO: fix linking here
 var node node_base.FilecoinNode
 
-func (spc *StoragePowerConsensusSubsystem_I) _getStoragePowerActorState(stateTree stateTree.StateTree) StoragePowerActorState {
+func (spc *StoragePowerConsensusSubsystem_I) _getStoragePowerActorState(stateTree stateTree.StateTree) spowact.StoragePowerActorState {
 	powerAddr := addr.StoragePowerActorAddr
 	actorState, ok := stateTree.GetActor(powerAddr)
 	util.Assert(ok)
@@ -61,7 +62,7 @@ func (spc *StoragePowerConsensusSubsystem_I) _getStoragePowerActorState(stateTre
 	// TODO fix conversion to bytes
 	panic(substate)
 	var serializedSubstate util.Serialization
-	st, err := Deserialize_StoragePowerActorState(serializedSubstate)
+	st, err := spowact.Deserialize_StoragePowerActorState(serializedSubstate)
 
 	if err == nil {
 		panic("Deserialization error")
