+++ src/systems/filecoin_mining/storage_mining/storage_miner_actor.md
@@ -6,12 +6,12 @@ title: Storage Miner Actor
 
 # `StorageMinerActor` interface
 
-{{< readfile file="storage_miner_actor.id" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/storage_miner/storage_miner_actor.id" code="true" lang="go" >}}
 
 # `StorageMinerActorState` implementation
 
-{{< readfile file="storage_miner_actor_state.go" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/storage_miner/storage_miner_actor_state.go" code="true" lang="go" >}}
 
 # `StorageMinerActorCode` implementation
 
-{{< readfile file="storage_miner_actor_code.go" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/storage_miner/storage_miner_actor_code.go" code="true" lang="go" >}}
