+++ src/systems/filecoin_vm/indices/indices.id
@@ -1,5 +1,5 @@
 import actor "github.com/filecoin-project/specs/systems/filecoin_vm/actor"
-import actor_util "github.com/filecoin-project/specs/systems/filecoin_vm/actor_util"
+import actor_util "github.com/filecoin-project/specs/actors/util"
 import block "github.com/filecoin-project/specs/systems/filecoin_blockchain/struct/block"
 import piece "github.com/filecoin-project/specs/systems/filecoin_files/piece"
 import sector "github.com/filecoin-project/specs/systems/filecoin_mining/sector"
