+++ src/systems/filecoin_vm/sysactors/account_actor.md
@@ -2,6 +2,5 @@
 title: AccountActor
 ---
 
-{{< readfile file="account_actor.id" code="true" lang="go" >}}
-
-{{< readfile file="account_actor.go" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/account/account_actor.id" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/account/account_actor.go" code="true" lang="go" >}}
