+++ src/systems/filecoin_vm/sysactors/cron_actor.md
@@ -2,6 +2,5 @@
 title: CronActor
 ---
 
-{{< readfile file="cron_actor.id" code="true" lang="go" >}}
-
-{{< readfile file="cron_actor.go" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/cron/cron_actor.id" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/cron/cron_actor.go" code="true" lang="go" >}}
