+++ src/systems/filecoin_vm/sysactors/init_actor.md
@@ -2,6 +2,5 @@
 title: InitActor
 ---
 
-{{< readfile file="init_actor.id" code="true" lang="go" >}}
-
-{{< readfile file="init_actor.go" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/init/init_actor.id" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/builtin/init/init_actor.go" code="true" lang="go" >}}
