--- src/systems/filecoin_vm/sysactors/sysactors_boilerplate.go
@@ -1,80 +1,0 @@
-package sysactors
-
-import (
-	ipld "github.com/filecoin-project/specs/libraries/ipld"
-	actor "github.com/filecoin-project/specs/systems/filecoin_vm/actor"
-	actor_util "github.com/filecoin-project/specs/systems/filecoin_vm/actor_util"
-	vmr "github.com/filecoin-project/specs/systems/filecoin_vm/runtime"
-	util "github.com/filecoin-project/specs/util"
-)
-
-////////////////////////////////////////////////////////////////////////////////
-// Boilerplate: System singleton actors
-//
-// This boilerplate should be essentially identical for all actors, and
-// conceptually belongs in the runtime/VM. It is only duplicated here as a
-// workaround due to the lack of generics support in Go.
-////////////////////////////////////////////////////////////////////////////////
-
-type InvocOutput = vmr.InvocOutput
-type Runtime = vmr.Runtime
-type Bytes = util.Bytes
-type Serialization = util.Serialization
-
-var Assert = util.Assert
-
-var CheckArgs = actor_util.CheckArgs
-var ArgPop = actor_util.ArgPop
-var ArgEnd = actor_util.ArgEnd
-
-var RT_ValidateImmediateCallerIsSignable = vmr.RT_ValidateImmediateCallerIsSignable
-
-////////////////////////////////////////////////////////////////////////////////
-// -- InitActor
-////////////////////////////////////////////////////////////////////////////////
-
-func _loadState(rt Runtime) (vmr.ActorStateHandle, InitActorState) {
-	h := rt.AcquireState()
-	stateCID := ipld.CID(h.Take())
-	var state InitActorState_I
-	if !rt.IpldGet(stateCID, &state) {
-		rt.AbortAPI("state not found")
-	}
-	return h, &state
-}
-func Release(rt Runtime, h vmr.ActorStateHandle, st InitActorState) {
-	checkCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
-	h.Release(checkCID)
-}
-func UpdateRelease(rt Runtime, h vmr.ActorStateHandle, st InitActorState) {
-	newCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
-	h.UpdateRelease(newCID)
-}
-func (st *InitActorState_I) CID() ipld.CID {
-	panic("TODO")
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// -- MultiSigActor
-////////////////////////////////////////////////////////////////////////////////
-
-func (a *MultiSigActorCode_I) State(rt Runtime) (vmr.ActorStateHandle, MultiSigActorState) {
-	h := rt.AcquireState()
-	stateCID := ipld.CID(h.Take())
-	var state MultiSigActorState_I
-	if !rt.IpldGet(stateCID, &state) {
-		rt.AbortAPI("state not found")
-	}
-	return h, &state
-}
-func Release_MultiSig(rt Runtime, h vmr.ActorStateHandle, st MultiSigActorState) {
-	checkCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
-	h.Release(checkCID)
-}
-func UpdateRelease_MultiSig(rt Runtime, h vmr.ActorStateHandle, st MultiSigActorState) {
-	newCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
-	h.UpdateRelease(newCID)
-}
-func (st *MultiSigActorState_I) CID() ipld.CID {
-	panic("TODO")
-}
